src/openclip/requirements.txt
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
appdirs==1.4.4
braceexpand==0.1.7
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dataclasses==0.6
docker-pycreds==0.4.0
eval_type_backport==0.2.0
ExifRead-nocycle==3.0.1
filelock==3.15.4
fire==0.5.0
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.6
idna==3.8
imageio==2.34.2
img2dataset==1.45.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==15.0.2
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-image==0.24.0
sentry-sdk==2.9.0
setproctitle==1.3.3
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sympy==1.13.2
termcolor==2.4.0
tifffile==2024.7.2
timm==1.0.9
torch==2.4.0
torchaudio==2.4.0+cu124
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
wandb==0.16.6
wcwidth==0.2.13
webdataset==0.2.100
wheel==0.43.0
